chore(common.tls): Refactor code into server and client
