+++ plugins/common/tls/server.conf
@@ -1,0 +1,19 @@
+  ## Used for TLS server certificate authentication
+  # tls_cert = "/path/to/certfile"
+  ## Used for TLS server certificate authentication
+  # tls_key = "/path/to/keyfile"
+  ## Password for encrypted key files
+  # tls_key_pwd = ""
+  ## CA certificates used for verifying client certificates
+  # tls_allowed_cacerts = []
+  ## List of ciphers to accept, by default all secure ciphers will be accepted
+  ## See https://pkg.go.dev/crypto/tls#pkg-constants for supported values.
+  ## Use "all", "secure" and "insecure" to add all support ciphers, secure
+  ## suites or insecure suites respectively.
+  # tls_cipher_suites = ["secure"]
+  ## Minimal TLS version to accept by the server
+  # tls_min_version = "TLS12"
+  ## Maximum TLS version to accept by the server
+  # tls_max_version = "TLS12"
+  ## Whitelist for certificate DNS names to accept
+  # tls_allowed_dns_names = []
